Turn smallQuery into a series of ES6 modules
Fixes #40
