+++ demo.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8">
         <title>smallQuery test page</title>
-        <script src="./smallquery.js"></script>
+        <script src="./src/smallquery.js" type="module"></script>
     </head>
     <body>
         <p>Open the console to play around with smallQuery</p>
