+++ spec/smallquery/append_spec.js
@@ -1,4 +1,4 @@
-"use strict";
+import $ from "../../src/smallquery.js";
 
 describe("the append function", function() {
     var $obj;
