+++ spec/smallquery/css_spec.js
@@ -1,4 +1,4 @@
-'use strict';
+import $ from "../../src/smallquery.js";
 
 describe("the css() function", function() {
     var $obj;
