+++ spec/smallquery/data_spec.js
@@ -1,4 +1,4 @@
-"use strict";
+import $ from "../../src/smallquery.js";
 
 describe("the data API", function() {
     var $obj;
