+++ spec/smallquery/siblings_spec.js
@@ -1,4 +1,4 @@
-"use strict";
+import $ from "../../src/smallquery.js";
 
 describe("the siblings function", function() {
     var $obj;
