+++ src/smallquery.js
@@ -1,0 +1,10 @@
+/*! smallquery Copyright  C 2019-2020 Matt Molyneaux */
+
+import smallQuery from "./core.js";
+
+import "./dom.js";
+import "./find.js";
+import "./css.js";
+import "./data.js";
+
+export default smallQuery;
